% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjG.R
\name{adjG}
\alias{adjG}
\title{Adjusted G matrix.}
\usage{
adjG(G, a)
}
\arguments{
\item{G}{A matrix of dimension \code{nObs} x \code{nEqs}.}

\item{a}{A tuning parameter.}
}
\value{
A matrix of dimension \code{nObs+1} x \code{nEqs}.
}
\description{
Adjusted G matrix.
}
\details{
...
}
