% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesELnew-package.R
\docType{package}
\name{bayesELnew}
\alias{bayesELnew}
\alias{bayesELnew-package}
\title{Bayesian Empirical Likelihood (new version)}
\description{
Bayesian Empirical Likelihood (new version)
}
\details{
Some details here.
}
