% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalEpsilons.R
\name{evalEpsilonsLS}
\alias{evalEpsilonsLS}
\title{Returns residuals of a location-scale model given the parameters.}
\usage{
evalEpsilonsLS(y, X, Z, beta, gamma, sig2)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{Matrix of dimension \code{nObs x nBet} of location covariates.}

\item{Z}{Matrix of dimension \code{nObs x nGam} of scale covariates.}

\item{beta}{Length-\code{nObs} vector of responses.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Length-\code{nObs} vector of residual epsilons.
}
\description{
Returns residuals of a location-scale model given the parameters.
}
\details{
...
}
