% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalPsos.smooth.R
\name{evalPsos.smooth}
\alias{evalPsos.smooth}
\title{Calculate partial sum of omegas with smoothing}
\usage{
evalPsos.smooth(ii, omegas, epsilons, s = 10)
}
\arguments{
\item{ii}{An integer indicator between 1 and the length of omegas/epsilons.}

\item{omegas}{Vector of probability weights}

\item{epsilons}{Vector of residuals.}
}
\value{
Partial sum of omegas according to residuals that are no larger than the ii-th residual.
}
\description{
Calculate partial sum of omegas with smoothing
}
\details{
...
}
