% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeights.R
\name{evalWeights}
\alias{evalWeights}
\title{Solve the inner loop optimization of an EL function.}
\usage{
evalWeights(y, X, deltas, omegas, beta)
}
\arguments{
\item{y}{Lenght-\code{nObs} vector of responses.}

\item{X}{\code{nObs x n Eqs} matrix of covariate matrix.}

\item{deltas}{Lenght-\code{nObs} vector of censoring indicators.}

\item{omegas}{Lenght-\code{nObs} vector, empirical distribution (probability vector).}
}
\value{
Length-\code{nObs} vector of weights for the weighted NR algorithm.
}
\description{
Solve the inner loop optimization of an EL function.
}
\details{
...
}
