% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeights.smooth.R
\name{evalWeights.smooth}
\alias{evalWeights.smooth}
\title{Calculate partial sum of omegas with smoothing}
\usage{
evalWeights.smooth(deltas, omegas, epsilons, s = 10)
}
\arguments{
\item{deltas}{Vector of censroing indicators.}

\item{omegas}{Vector of probability weights}

\item{epsilons}{Vector of residuals.}
}
\value{
Partial sum of omegas according to residuals that are no larger than the ii-th residual.
}
\description{
Calculate partial sum of omegas with smoothing
}
\details{
...
}
