% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexEL-package.R
\docType{package}
\name{flexEL}
\alias{flexEL}
\title{Fast and flexible empirical likelihood calculations.}
\description{
A flexible framework for Empirical Likelihood methods for regression problems. The bulk of computations are done in C++ for speed considerations.  The C++ code is exposed as a header-only library.  The package provides mean and quantile regression with (optionally) right-censored responses using location and location-scale models, but users can added in any regression models as C++ header files or use the R interface directly to solve their problems.
}
\keyword{internal}
