% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GenEL.R
\name{GenEL}
\alias{GenEL}
\title{R6 class for EL with general moment specification.}
\description{
R6 class for EL with general moment specification.

R6 class for EL with general moment specification.
}
\details{
A general EL object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenEL$new()}}
\item \href{#method-set_max_iter}{\code{GenEL$set_max_iter()}}
\item \href{#method-set_rel_tol}{\code{GenEL$set_rel_tol()}}
\item \href{#method-set_supp_adj}{\code{GenEL$set_supp_adj()}}
\item \href{#method-set_lambda0}{\code{GenEL$set_lambda0()}}
\item \href{#method-lambda_nr}{\code{GenEL$lambda_nr()}}
\item \href{#method-omega_hat}{\code{GenEL$omega_hat()}}
\item \href{#method-logel}{\code{GenEL$logel()}}
\item \href{#method-clone}{\code{GenEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Check the dimension of G is what is expected.


Access or reset the value of .G.


Create a new GenEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$new(n_obs, n_eqs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_obs}}{Number of observations.}

\item{\code{n_eqs}}{Number of (moment constraint) equations.}

\item{\code{G}}{A numeric matrix.}

\item{\code{value}}{Missing or a numeric matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_iter"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_iter}{}}}
\subsection{Method \code{set_max_iter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_max_iter(max_iter = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rel_tol"></a>}}
\if{latex}{\out{\hypertarget{method-set_rel_tol}{}}}
\subsection{Method \code{set_rel_tol()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_rel_tol(rel_tol = 1e-07)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_supp_adj"></a>}}
\if{latex}{\out{\hypertarget{method-set_supp_adj}{}}}
\subsection{Method \code{set_supp_adj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_supp_adj(supp_adj = FALSE, a = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_lambda0"></a>}}
\if{latex}{\out{\hypertarget{method-set_lambda0}{}}}
\subsection{Method \code{set_lambda0()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_lambda0(lambda0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda_nr"></a>}}
\if{latex}{\out{\hypertarget{method-lambda_nr}{}}}
\subsection{Method \code{lambda_nr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$lambda_nr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega_hat"></a>}}
\if{latex}{\out{\hypertarget{method-omega_hat}{}}}
\subsection{Method \code{omega_hat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$omega_hat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
