% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeights.R
\name{getWeights}
\alias{getWeights}
\title{Solve the inner loop optimization of an EL function.}
\usage{
getWeights(deltas, omegas, epsilons)
}
\arguments{
\item{deltas}{Lenght-\code{nObs} vector of censoring indicators.}

\item{omegas}{Lenght-\code{nObs} vector, empirical distribution (probability vector).}
}
\value{
Length-\code{nObs} vector of weights for the weighted NR algorithm.
}
\description{
Solve the inner loop optimization of an EL function.
}
\details{
...
}
