% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm.R
\name{hlm}
\alias{hlm}
\alias{hlm_loglik}
\alias{hlm_loglik}
\title{Fit the normal heteroscedastic linear model with censoring.}
\usage{
hlm(y, delta, X, W, max_iter = 1000, rel_tol = 1e-05, multi_cycle = FALSE)

hlm_loglik(beta, gamma, y, delta, X, W)
}
\arguments{
\item{y}{Response vector.}

\item{delta}{Censoring indicator vector (1: observed, 0: censored).}

\item{X}{Covariate matrix for the mean term.  Must include the intercept if required.}

\item{W}{Covariate matrix for the variance term.  Must include the intercept if required.}

\item{max_iter}{Maximum number of iterations of ECM algorithm (see Details).}

\item{rel_tol}{Maximum relative parameter difference between successive iterations (see Details).}

\item{multi_cycle}{Whether or not to compute the E-step after both CM steps, or only after the second one.  The former converges in fewer steps, but each step is longer.}
}
\value{
For \code{hlm_loglik}, the loglikelihood function.  For \code{hlm}, a list with elements
\describe{
  \item{\code{coef}}{The vector \code{(beta_hat, gamma_h)} giving the MLE estimates for the model.}
  \item{\code{vcov}}{The estimated variance for the MLE.}
  \item{\code{loglik}}{The value of the loglikelihood function evaluated at the MLE.}
  \item{\code{niter}}{The number of ECM iterations.}
}
}
\description{
Fit the normal heteroscedastic linear model with censoring.
}
\details{
The HLM model is given by
\preformatted{
z_i = x_i'beta + exp(.5 w_i'gamma) * eps_i,    eps_i ~iid N(0,1),

y_i = min(z_i, c_i),   delta_i = 1[z_i <= c_i].
}
Maximum likelihood estimates of \code{(beta, gamma)} are obtained by an ECM method extending Smyth (1989).  That is, without censoring the parameters can be sequentially updated using \code{lm} for \code{beta}, and \code{glm(family = "Gamma", link = "log")} for \code{gamma}.

The function \code{hlm} does the fitting, and \code{hlm_loglik} computes the loglikelihood function.
}
