% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.smooth.R
\name{ind.smooth}
\alias{ind.smooth}
\title{Solve the inner loop optimization of an EL function.}
\usage{
ind.smooth(x, s = 10)
}
\arguments{
\item{x}{Scalar or vector of real values.}

\item{s}{Scalar or vector of the same length as \code{x}, parameter of the smoothed indicator function.}
}
\value{
Scalar or vector of the same length as \code{x} where the values are between 0 and 1 evaluated by the smoothed indicator function.
}
\description{
Solve the inner loop optimization of an EL function.
}
\details{
...
}
