% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind1.smooth.R
\name{ind1.smooth}
\alias{ind1.smooth}
\title{First derivative of smoothen indicator function}
\usage{
ind1.smooth(x, s = 10)
}
\arguments{
\item{x}{Scalar or vector of real values.}

\item{s}{Scalar or vector of the same length as \code{x}, parameter of the smoothed indicator function.}
}
\value{
Scalar or vector of the same length as \code{x} where the values are between 0 and 1 evaluated by the smoothed indicator function.
}
\description{
First derivative of smoothen indicator function
}
\details{
...
}
