% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaNR.R
\name{lambdaNR}
\alias{lambdaNR}
\title{Solves the inner optimization problem of an EL maximization.}
\usage{
lambdaNR(G, max_iter = 100, rel_tol = 1e-07, support = FALSE, verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}

\item{max_iter}{Integer scalar for maximum number of Newton-Raphson steps.}

\item{rel_tol}{Positive scalar for relative tolerance of Newton-Raphson steps.}

\item{verbose}{Logical scalar for whether to display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Vector of length \code{n_eq} corresponding to the solution of the optimization problem.
}
\description{
Solves the inner optimization problem of an EL maximization.
}
\examples{
n <- 20
p <- 8
G <- matrix(rnorm(n*p),n,p) # randomly generated G
max_iter <- 100
rel_tol <- 1e-5

# without support correction
lambdaNR(G = G,
         max_iter = max_iter, rel_tol = rel_tol, support = FALSE, 
         verbose = FALSE)

# with support correction
lambdaNR(G = G,
         max_iter = max_iter, rel_tol = rel_tol, support = TRUE, 
         verbose = FALSE)
}
