% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaNR_cens.R
\name{lambdaNR_cens}
\alias{lambdaNR_cens}
\title{Solve the inner loop optimization of a censored EL function}
\usage{
lambdaNR_cens(G, weights, max_iter = 100, rel_tol = 1e-07,
  support = FALSE, verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}

\item{weights}{Length-\code{nObs} vector of weights for weighted Newton-Raphson algorithm.}

\item{max_iter}{Integer scalar for maximum number of Newton-Raphson steps.}

\item{rel_tol}{Positive scalar for relative tolerance of Newton-Raphson steps.}

\item{verbose}{Logical scalar for whether to display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Length-\code{nEq} vector corresponding to the solution of the optimization problem.
}
\description{
Solve the inner loop optimization of a censored EL function
}
\details{
The inner-loop optimization of EL is ...
}
