% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{Calculates log empirical likelihood given fully observed or right-censored observations.}
\usage{
logEL(omega, eps, delta, support = FALSE)
}
\arguments{
\item{omega}{Length-\code{nObs} probability vector.}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{delta}{Length-\code{nObs} vector of censoring indicators.}
}
\value{
Log empirical likelihood
}
\description{
Calculates log empirical likelihood given fully observed or right-censored observations.
}
\examples{
# no censoring
n <- 20
omegas <- abs(rnorm(n))
omegas <- omegas/sum(omegas) # create a random probability vector
logEL(omegas)

# right censoring
n <- 20 # total number of observations
k <- 4 # number of censored observations
idx <- sample(n,k) # indices of censored observations
deltas <- rep(1,n)
deltas[idx] <- 0 # censoring indicators
omegas <- abs(rnorm(n)) 
omegas <- omegas/sum(omegas) # create a random probability vector
epsilons <- rnorm(n) # time to event for each observation
logEL(omegas,epsilons,deltas)
}
