% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.smooth.R
\name{logEL.smooth}
\alias{logEL.smooth}
\title{Calculate the smoothed log empirical likelihood under censoring}
\usage{
logEL.smooth(omegas, epsilons, deltas, s = 10)
}
\arguments{
\item{omegas}{Vector of probability weights}

\item{epsilons}{Vector of residuals.}
}
\value{
A scalar of log empirical likelihood
}
\description{
Calculate the smoothed log empirical likelihood under censoring
}
\details{
...
}
