% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL_cens.R
\name{logEL_cens}
\alias{logEL_cens}
\title{Calculates log empirical likelihood for right-censored observations with optional continuity correction.}
\usage{
logEL_cens(omega, eps, delta, sp = 10, support = FALSE)
}
\arguments{
\item{omega}{Length-\code{nObs} probability vector.}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{delta}{Length-\code{nObs} vector of censoring indicators.}

\item{sp}{A positive scalar as smoothing parameter.}

\item{support}{A boolean indicating whether support correction should be conducted.}
}
\value{
Log empirical likelihood with continuity correction.
}
\description{
Calculates log empirical likelihood for right-censored observations with optional continuity correction.
}
