% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantReg_logEL.R
\name{logEL_QuantReg}
\alias{logEL_QuantReg}
\title{log empirical likelihood for quantile regression}
\usage{
logEL_QuantReg(X, y, nObs, nEqs, beta, alpha, lambda0, max_iter = 100,
  eps = 1e-07)
}
\arguments{
\item{X}{\code{nVars x nObs} matrix of constraints.}

\item{y}{Length-\code{nObs} vector of response values.}

\item{nObs}{Number of observations.}

\item{nEqs}{Number of estimating equations.}

\item{beta}{Length-\code{nVars} vector of coefficients.}

\item{alpha}{Quantile level, a calar between 0 and 1.}

\item{lambda0}{Length-\code{nEqs} vector of initial value of lambda.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{eps}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood for quantile regression
}
\details{
...
}
