% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.logel.R
\name{mr.logel}
\alias{mr.logel}
\title{log empirical likelihood for mean regression}
\usage{
mr.logel(y, X, beta, max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nEqs x nObs} matrix of constraints.}

\item{beta}{Length-\code{nEqs} vector of coefficients.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood for mean regression
}
\details{
...
}
