% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.post.R
\name{mr.post}
\alias{mr.post}
\title{Posterior sampler for mean regression (location model)}
\usage{
mr.post(y, X, nsamples, nburn, BetaInit, Sigs, max_iter = 100,
  rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nEqs} matrix of constraints.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{BetaInit}{\code{nEqs x numBeta} matrix of initial value for the chain. Each column is a beta vector. If a vector is passed to this function, it will be converted to a matrix with only one column.}

\item{Sigs}{\code{nEqs x numBeta} matrix of tuning parameters.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for mean regression (location model)
}
\details{
...
}
