% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_cens.neglogEL.smooth.R
\name{mr_cens.neglogEL.smooth}
\alias{mr_cens.neglogEL.smooth}
\title{Calculate the negative log EL with smoothing for mean regression under censoring}
\usage{
mr_cens.neglogEL.smooth(y, X, delta, beta, sp = 10)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates in the location function, where \code{nObs} is the number of observations and \code{nBet} is the dimension of coefficients in the location function.}

\item{delta}{Length-\code{nObs} vector of censoring indicators.}

\item{beta}{Length-\code{nBet} vector of coefficients in mean regression model.}

\item{sp}{Positive scalar as smoothing parameter.}
}
\value{
Scalar of negative smoothed log EL.
}
\description{
Calculate the negative log EL with smoothing for mean regression under censoring
}
\details{
...
}
