% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_cens.post_adapt.R
\name{mr_cens.post_adapt}
\alias{mr_cens.post_adapt}
\title{Posterior sampler for mean regression (location model)}
\usage{
mr_cens.post_adapt(y, X, delta, nsamp, nburn, betaInit, mwgSd, doMcmc, DoAdapt,
  max_iter = 100, rel_tol = 1e-07, abs_tol = 0.001)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates in the location function, where \code{nObs} is the number of observations and \code{nBet} is the dimension of coefficients in the location function.}

\item{delta}{Length-\code{nObs} vector of censoring indicators.}

\item{nsamp}{Positive integer of number of samples to obtain.}

\item{nburn}{Number of iterations to discard before saving the samples.}

\item{betaInit}{Length-\code{nEqs} vector of initial values for beta.}

\item{mwgSd}{Length-\code{nEqs} vector of tuning parameters for the Metropolis-Within-Gibbs sampler.}

\item{doMcmc}{Length-\code{nEqs} vector of 1 and 0s, where 1 indicates to update the corresponding entry of beta.}

\item{max_iter}{Integer scalar for maximum number of Newton-Raphson steps.}

\item{rel_tol}{Positive scalar for relative tolerance of Newton-Raphson steps.}
}
\value{
\code{nEqs x nsamp} matrix of posterior samples.
}
\description{
Posterior sampler for mean regression (location model)
}
\details{
...
}
