% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_dGdt.R
\name{mr_dGdt}
\alias{mr_dGdt}
\title{Calculate derivative of G w.r.t theta for mean regression model}
\usage{
mr_dGdt(y, X, beta)
}
\value{
A list of length \code{n_obs}.
}
\description{
Calculate derivative of G w.r.t theta for mean regression model
}
