% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_evalG.R
\name{mr_evalG}
\alias{mr_evalG}
\title{Evaluate the G matrix for mean regression (location model)}
\usage{
mr_evalG(y, X, beta)
}
\arguments{
\item{y}{Length-\code{n_obs} vector of responses.}

\item{X}{\code{n_obs x n_eqs} matrix of covariates.}

\item{beta}{Length-\code{n_eqs} vector of coefficients in location model.}
}
\value{
G matrix for location mean regression model.
}
\description{
Evaluate the G matrix for mean regression (location model)
}
\details{
...
}
