% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls.evalG.R
\name{mrls.evalG}
\alias{mrls.evalG}
\title{Evaluate the G matrix for mean regression (location-scale model)}
\usage{
mrls.evalG(y, X, Z, beta, gamma, sig2)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates.}

\item{Z}{\code{nObs x nGam} matrix of covariates.}

\item{beta}{Length-\code{nBet} vector of coefficients in location model.}

\item{gamma}{Length-\code{nGam} vector of coefficients in location model.}

\item{sig2}{A scalar where sqrt(sig2) is the scale parameter for the error term.}
}
\value{
G matrix for location-scale mean regression model.
}
\description{
Evaluate the G matrix for mean regression (location-scale model)
}
\details{
Returns the G matrix for location-scale mean regression model.
}
