% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanRegLS.R
\name{mrls.logel}
\alias{mrls.logel}
\title{log empirical likelihood for mean regression with location scale model}
\usage{
mrls.logel(y, X, Z, beta, gamma, max_iter = 100, eps = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nVars x nObs} matrix of constraints.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{eps}{Relative tolerance of Newton-Raphson convergence.}

\item{nObs}{Number of observations.}

\item{nEqs}{Number of estimating equations.}

\item{theta}{Length-\code{nVars} vector of coefficients.}

\item{lambda0}{Length-\code{nEqs} vector of initial value of lambda.}
}
\value{
Log empirical likelihood of the input theta
}
\description{
log empirical likelihood for mean regression with location scale model
}
\details{
...
}
