% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls.post.R
\name{mrls.post}
\alias{mrls.post}
\title{Posterior sampler for mean regression (location-scale model)}
\usage{
mrls.post(y, X, Z, nsamples, nburn, betaInit, gammaInit, sigs, max_iter = 100,
  rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nBet} matrix of location function covariates.}

\item{Z}{\code{nObs x nGam} matrix of scale function covariates.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{betaInit}{Length-\code{nBet} vector of initial value for the chain.}

\item{gammaInit}{Length-\code{nGam} vector of initial value for the chain.}

\item{sigs}{Length-\code{nEqs} vector of tuning parameters.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{betalen}{A scalar of the length of the location model parameter.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for mean regression (location-scale model)
}
\details{
...
}
