% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls_cens.post_adapt.R
\name{mrls_cens.post_adapt}
\alias{mrls_cens.post_adapt}
\title{Posterior sampler for mean regression (location-scale model)}
\usage{
mrls_cens.post_adapt(y, X, Z, deltas, nsamples, nburn, betaInit, gammaInit,
  sig2Init, mwgSd, rvDoMcmc, doAdapt, max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nEqs} matrix of constraints.}

\item{deltas}{Length-\code{nObs} vector of censoring indicators.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{betaInit}{\code{nEqs} vector of initial value for the chain.}

\item{mwgSd}{Length-\code{nObs} vector of tuning parameters.}

\item{rvDoMcmc}{Length-\code{nEqs} vector of 1 and 0s, where 1 indicates to update the corresponding entry of beta.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for mean regression (location-scale model)
}
\details{
...
}
