% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls_dGdt.R
\name{mrls_dGdt}
\alias{mrls_dGdt}
\title{Calculate derivative of G w.r.t theta for location-scale mean regression model}
\usage{
mrls_dGdt(y, X, Z, beta, gamma, sig2)
}
\value{
A list of length \code{n_obs}.
}
\description{
Calculate derivative of G w.r.t theta for location-scale mean regression model
}
