% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls_evalG.R
\name{mrls_evalG}
\alias{mrls_evalG}
\title{Evaluate the G matrix for mean regression (location-scale model)}
\usage{
mrls_evalG(y, X, Z, beta, gamma, sig2)
}
\arguments{
\item{y}{Length-\code{n_obs} vector of responses.}

\item{X}{\code{n_obs x n_bet} matrix of covariates.}

\item{Z}{\code{n_obs x n_gam} matrix of covariates.}

\item{beta}{Length-\code{n_bet} vector of coefficients in location model.}

\item{gamma}{Length-\code{n_gam} vector of coefficients in location model.}

\item{sig2}{A scalar where sqrt(sig2) is the scale parameter for the error term.}
}
\value{
G matrix for location-scale mean regression model.
}
\description{
Evaluate the G matrix for mean regression (location-scale model)
}
\details{
Returns the G matrix for location-scale mean regression model.
}
