% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.hat.EM.smooth.R
\name{omega.hat.EM.smooth}
\alias{omega.hat.EM.smooth}
\title{Returns the empirical distribution, omegas, from empirical likelihood (EL) maximization.}
\usage{
omega.hat.EM.smooth(G, deltas, epsilons, sp = 10, max_iter = 100,
  rel_tol = 1e-05, abs_tol = 0.001, support = FALSE, verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqns} matrix of constraints.}

\item{deltas}{Length-\code{nObs} vector of censor indicators, omit if no censoring.}

\item{epsilons}{Length-\code{nObs} vector of residuals, omit if no censoring.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{abs_tol}{Absolute tolerance of Newton-Raphson convergence.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Length-\code{nEq} vector for the resulting empirical distribution, omegas, if the optimization algorithm converged; 1/nObs if did not converge.
}
\description{
Returns the empirical distribution, omegas, from empirical likelihood (EL) maximization.
}
\details{
The inner-loop optimization of EL is ...
}
