% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.hat.R
\name{omega.hat}
\alias{omega.hat}
\title{Returns the omegas, i.e., the empirical probabilities.}
\usage{
omega.hat(G, deltas, epsilons, max_iter = 100, rel_tol = 1e-07,
  verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqns} matrix of constraints.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}

\item{eps}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
Length-\code{nEq} vector corresponding to the solution of the optimization problem.
}
\description{
Returns the omegas, i.e., the empirical probabilities.
}
\details{
The inner-loop optimization of EL is ...
}
