% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.evalG.R
\name{qr.evalG}
\alias{qr.evalG}
\title{Evaluate the G matrix for quantile regression (location model)}
\usage{
\method{qr}{evalG}(y, X, alphas, Beta)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nEqs} matrix of covariates.}

\item{alphas}{a vector of quantile levels.}

\item{Beta}{\code{nEqs x nQts} matrix, each column is a vector of coefficients in location model.}
}
\value{
G matrix for location quantile regression model.
}
\description{
Evaluate the G matrix for quantile regression (location model)
}
\details{
...
}
