% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.logel.R
\name{qr.logel}
\alias{qr.logel}
\title{log empirical likelihood for quantile regression}
\usage{
\method{qr}{logel}(y, X, alpha, beta, max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nVars x nObs} matrix of constraints.}

\item{alpha}{Quantile level, a calar between 0 and 1.}

\item{beta}{Length-\code{nVars} vector of coefficients.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{nObs}{Number of observations.}

\item{nEqs}{Number of estimating equations.}

\item{lambda0}{Length-\code{nEqs} vector of initial value of lambda.}

\item{eps}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood for quantile regression
}
\details{
...
}
