% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.post_adapt.R
\name{qr.post_adapt}
\alias{qr.post_adapt}
\title{Posterior sampler for quantile regression (location model)}
\usage{
\method{qr}{post_adapt}(y, X, alphas, nsamples, nburn, betaInit, mwgSd,
  rvDoMcmc, max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nEqs} matrix of constraints.}

\item{alphas}{a vector of quantile levels.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{betaInit}{\code{nEqs} vector of initial value for the chain.}

\item{mwgSd}{Length-\code{nObs} vector of tuning parameters.}

\item{rvDoMcmc}{Length-\code{nEqs} vector of 1 and 0s, where 1 indicates to update the corresponding entry of beta.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for quantile regression (location model)
}
\details{
...
}
