% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_cens.post.R
\name{qr_cens.post}
\alias{qr_cens.post}
\title{Posterior sampler for quantile regression with censoring (location model)}
\usage{
qr_cens.post(y, X, deltas, alpha, nsamples, nburn, BetaInit, Sigs, RvDoMcmc,
  max_iter = 100, rel_tol = 1e-07, abs_tol = 0.001)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nEqs} matrix of constraints.}

\item{deltas}{Length-\code{nObs} vector of censoring indicators.}

\item{alpha}{a scalar of quantile level (TODO: multiple quantile levels).}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{BetaInit}{\code{nEqs x numBeta} matrix of initial value for the chain. Each column is a beta vector. If a vector is passed to this function, it will be converted to a matrix with only one column.}

\item{Sigs}{\code{nEqs x numBeta} matrix of tuning parameters.}

\item{RvDoMcmc}{\code{nEqs x numBeta} matrix of indicators (1 or 0) for whether or not updating the corresponding parameter.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for quantile regression with censoring (location model)
}
\details{
...
}
