% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_dGdt.R
\name{qr_dGdt}
\alias{qr_dGdt}
\title{Calculate derivative of G w.r.t theta for quantile regression model}
\usage{
qr_dGdt(y, X, Beta, alpha, sp)
}
\arguments{
\item{y}{A numeric vector of resoponses of length \code{n_obs} where \code{n_obs}
is the number of observations.}

\item{X}{A numeric matrix of covariates of dimension \code{n_obs} x \code{n_bet}
where \code{n_obs} is the number of observations and \code{b_bet} is the number
of coefficients (length of \code{beta}).}

\item{beta}{A numeric vector of coefficients.}
}
\value{
A list of length \code{n_obs}.
}
\description{
Calculate derivative of G w.r.t theta for quantile regression model
}
