% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls.evalG.R
\name{qrls.evalG}
\alias{qrls.evalG}
\title{Evaluate the G matrix for quantile regression location-scale model}
\usage{
qrls.evalG(y, X, Z, alphas, Beta, Gamma, Sig2, Nu)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates.}

\item{Z}{\code{nObs x nGam} matrix of covariates.}

\item{alphas}{a vector of quantile levels.}

\item{Beta}{\code{nBet x nQts} matrix, each column is a vector of coefficients in location function.}

\item{Gamma}{\code{nGam x nQts} matrix, each column is a vector of coefficients in scale function.}

\item{Nu}{Length-\code{numNu} vector of initial value for the chain.}
}
\value{
G matrix for location-scale quantile regression model.
}
\description{
Evaluate the G matrix for quantile regression location-scale model
}
