% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls.post.R
\name{qrls.post}
\alias{qrls.post}
\title{Posterior sampler for quantile regression (location-scale model)}
\usage{
qrls.post(y, X, Z, alphas, nsamples, nburn, BetaInit, GammaInit, Sig2Init,
  NuInit, Sigs, RvDoMcmc, max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nBet} matrix of location function covariates.}

\item{Z}{\code{nObs x nGam} matrix of scale function covariates.}

\item{alphas}{a vector of quantile levels.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{BetaInit}{\code{nBet x numBeta} matrix of initial value for the chain.}

\item{GammaInit}{\code{nGam x numGamma} matrix of initial value for the chain.}

\item{NuInit}{Length-\code{numNu} vector of initial value for the chain.}

\item{Sigs}{Length-\code{nEqs} vector of tuning parameters.}

\item{RvDoMcmc}{\code{nEqs x numBeta} matrix of indicators (1 or 0) for whether or not updating the corresponding parameter.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for quantile regression (location-scale model)
}
\details{
...
}
