% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls.post_adapt.R
\name{qrls.post_adapt}
\alias{qrls.post_adapt}
\title{Adaptive posterior sampler for qunatile regression (location-scale model)}
\usage{
qrls.post_adapt(y, X, Z, alpha, nsamples, nburn, betaInit, gammaInit, sig2Init,
  nuInit, mwgSd, RvDoMcmc, max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nBet} matrix of location function covariates.}

\item{Z}{\code{nObs x nGam} matrix of scale function covariates.}

\item{alpha}{a scalar of quantile level (TODO: multiple quantile levels).}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{betaInit}{\code{nBet} vector of initial value for the parameters in the linear location function.}

\item{gammaInit}{\code{nGam} vector of initial value for the parameters in the exponential scale function.}

\item{sig2Init}{A scalar of initial value for scale parameter.}

\item{nuInit}{A scalar of initial value for the quantile value of the error term.}

\item{mwgSd}{Length-\code{nEqs} vector of tuning parameters.}

\item{RvDoMcmc}{\code{nEqs x numBeta} matrix of indicators (1 or 0) for whether or not updating the corresponding parameter.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Adaptive posterior sampler for qunatile regression (location-scale model)
}
\details{
...
}
