% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls_cens.neglogEL.smooth.R
\name{qrls_cens.neglogEL.smooth}
\alias{qrls_cens.neglogEL.smooth}
\title{Calculate the negative log EL with smoothing for quantile regression location-scale model under censoring}
\usage{
qrls_cens.neglogEL.smooth(y, X, Z, deltas, tau, theta, sp = 10)
}
\arguments{
\item{omegas}{Vector of probability weights}

\item{epsilons}{Vector of residuals.}

\item{ii}{An integer indicator between 1 and the length of omegas/epsilons.}
}
\value{
Partial sum of omegas according to residuals that are no larger than the ii-th residual.
}
\description{
Calculate the negative log EL with smoothing for quantile regression location-scale model under censoring
}
\details{
...
}
