% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls_evalG.R
\name{qrls_evalG}
\alias{qrls_evalG}
\title{Evaluate the G matrix for a location-scale quantile regression model.}
\usage{
qrls_evalG(y, X, Z, alphas, Beta, Gamma, Sig2, Nu)
}
\arguments{
\item{y}{Response vector of length \code{n_obs}.}

\item{X}{Location covariate matrix of size \verb{n_obs x n_bet}.}

\item{Z}{Scale covariate matrix of size \verb{n_obs x n_gam}.}

\item{alphas}{Vector of \code{n_qts} quantile levels.}

\item{Beta}{Matrix of \verb{n_bet x n_qts} location coefficients.}

\item{Gamma}{Matrix of \verb{n_gam x n_qts} scale coefficients.}

\item{Nu}{Vector of \code{n_nu} initial values for the chain.}
}
\value{
G matrix of size `for location-scale quantile regression model.
}
\description{
Evaluate the G matrix for a location-scale quantile regression model.
}
\details{
...
}
\examples{
n <- 20
p <- 5
q <- 3
X <- replicate(p, rnorm(n))
Z <- replicate(q, rnorm(n))
alpha <- runif(1)
beta <- rnorm(p)
gamma <- rnorm(q)
nu <- rnorm(1)
sig2 <- 1.5
y <- c(X \%*\% beta + sqrt(sig2)*exp(Z \%*\% gamma)*rnorm(n)) # with multiplicative N(0,1) error
qrls_evalG(y,X,Z,alpha,beta,gamma,sig2,nu)
}
