% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantReg_post.R
\name{QuantReg_post}
\alias{QuantReg_post}
\title{log empirical likelihood for mean regression}
\usage{
QuantReg_post(nObs, nEqs, y, X, alpha, lambda0, nsamples, nburn, betaInit, sigs,
  max_iter = 100, eps = 1e-07)
}
\arguments{
\item{nObs}{Number of observations.}

\item{nEqs}{Number of estimating equations.}

\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nVars x nObs} matrix of constraints.}

\item{lambda0}{Length-\code{nEqs} vector of initial value of lambda.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{eps}{Relative tolerance of Newton-Raphson convergence.}

\item{beta}{Length-\code{nVars} vector of coefficients.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood for mean regression
}
\details{
...
}
