% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjG.R
\name{adjG}
\alias{adjG}
\title{Adjusted G matrix.}
\usage{
adjG(G, a)
}
\arguments{
\item{G}{A matrix of dimension \verb{n_obs x n_eq}.}

\item{a}{A tuning parameter.}
}
\value{
A matrix of dimension \verb{(n_obs+1) x }n_eq`.
}
\description{
Adjusted G matrix.
}
\details{
...
}
