% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_CensEL.R
\name{CensEL}
\alias{CensEL}
\title{R6 class for EL regression where response is under right censoring.}
\description{
R6 class for EL regression where response is under right censoring.

R6 class for EL regression where response is under right censoring.
}
\details{
A general EL object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CensEL$new()}}
\item \href{#method-set_supp_adj}{\code{CensEL$set_supp_adj()}}
\item \href{#method-set_opts}{\code{CensEL$set_opts()}}
\item \href{#method-eval_weights}{\code{CensEL$eval_weights()}}
\item \href{#method-omega_hat}{\code{CensEL$omega_hat()}}
\item \href{#method-logel}{\code{CensEL$logel()}}
\item \href{#method-clone}{\code{CensEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Check the dimension of G is what is expected.


Access or reset the initial value of lambda.


Access or reset the value of the maximum number of iterations for the Newton-Raphson algorithm.


Access or reset the value of relative tolerance controlling the accuracy at convergence of the Newton-Raphson algorithm.


Access or reset the value of the maximum number of iterations for the EM algorithm.


Access or reset the value of absolute tolerance controlling the accuracy at convergence of the EM algorithm.


Access or reset the support correction flag.


Access or reset the value of support corection factor.


Create a new CensEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$new(n_obs, n_eqs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_obs}}{Number of observations.}

\item{\code{n_eqs}}{Number of (moment constraint) equations.}

\item{\code{G}}{A numeric matrix.}

\item{\code{value}}{Missing or a vector of length \code{n_eqs}.}

\item{\code{value}}{Missing or a positive integer (default to 100).}

\item{\code{value}}{Missing or a small positive number (default to 1e-7).}

\item{\code{value}}{Missing or a positive integer (default to 100).}

\item{\code{value}}{Missing or a small positive number (default to 1e-3).}

\item{\code{value}}{Missing or a boolean indicating whether to conduct support correction or not.}

\item{\code{value}}{Missing or a scalar. Defaults to \code{max(1.0, log(n_obs)/2)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{CensEL} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_supp_adj"></a>}}
\if{latex}{\out{\hypertarget{method-set_supp_adj}{}}}
\subsection{Method \code{set_supp_adj()}}{
Set the support correction flag and support correction factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$set_supp_adj(supp_adj = FALSE, supp_adj_a = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{supp_adj}}{A boolean indicating whether to conduct support correction or not.}

\item{\code{supp_adj_a}}{Support adjustment factor. Defaults to \code{max(1.0, log(n_obs)/2)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_opts"></a>}}
\if{latex}{\out{\hypertarget{method-set_opts}{}}}
\subsection{Method \code{set_opts()}}{
Set more than one options together.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$set_opts(
  max_iter_nr = 100,
  rel_tol = 1e-07,
  max_iter_em = 100,
  abs_tol = 0.001,
  lambda0 = rep(0, CensEL_get_n_eqs(private$.CEL)),
  supp_adj = FALSE,
  supp_adj_a = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iter_nr}}{A positive integer controlling the maximum number of iterations for the Newton-Raphson algorithm.}

\item{\code{rel_tol}}{A small positive number controlling accuracy at convergence for the Newton-Raphson algorithm.}

\item{\code{max_iter_em}}{A positive integer controlling the maximum number of iterations for the EM algorithm.}

\item{\code{abs_tol}}{A small positive number controlling accuracy at convergence for the EM algorithm.}

\item{\code{lambda0}}{Initialization vector of size \code{n_eqs}.}

\item{\code{supp_adj}}{A boolean indicating whether to conduct support correction or not.}

\item{\code{supp_adj_a}}{Support adjustment factor. Defaults to \code{max(1.0, log(n_obs)/2)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_weights"></a>}}
\if{latex}{\out{\hypertarget{method-eval_weights}{}}}
\subsection{Method \code{eval_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$eval_weights(delta, epsilon, omega)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega_hat"></a>}}
\if{latex}{\out{\hypertarget{method-omega_hat}{}}}
\subsection{Method \code{omega_hat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$omega_hat(G, delta, epsilon)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
Calculate the log empirical likelihood base on the given G matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$logel(G, delta, epsilon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{A matrix of dimension \verb{n_eqs x n_obs}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scalar.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CensEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
