% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWeights.R
\name{evalWeights}
\alias{evalWeights}
\title{Calculate the weights for the weighted (censored) EL problem}
\usage{
evalWeights(deltas, omegas, epsilons)
}
\arguments{
\item{deltas}{Lenght-\code{nObs} vector of censoring indicators.}

\item{omegas}{Lenght-\code{nObs} vector, empirical distribution (probability vector).}
}
\value{
Length-\code{nObs} vector of weights for the weighted NR algorithm.
}
\description{
Calculate the weights for the weighted (censored) EL problem
}
\details{
...
}
