% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexEL-package.R
\docType{package}
\name{flexEL}
\alias{flexEL}
\alias{flexEL-package}
\title{Fast and flexible empirical likelihood calculations.}
\description{
Fast and flexible empirical likelihood calculations.
}
\keyword{internal}
