% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GenEL.R
\name{GenEL}
\alias{GenEL}
\title{R6 class for EL with general moment specification.}
\description{
R6 class for EL with general moment specification.

R6 class for EL with general moment specification.
}
\details{
A general EL object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenEL$new()}}
\item \href{#method-set_supp_adj}{\code{GenEL$set_supp_adj()}}
\item \href{#method-set_opts}{\code{GenEL$set_opts()}}
\item \href{#method-lambda_nr}{\code{GenEL$lambda_nr()}}
\item \href{#method-omega_hat}{\code{GenEL$omega_hat()}}
\item \href{#method-logel}{\code{GenEL$logel()}}
\item \href{#method-logel_grad}{\code{GenEL$logel_grad()}}
\item \href{#method-clone}{\code{GenEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Check the dimension of G is what is expected.


Access or reset the value of .max_iter.


Access or reset the value of .rel_tol.


Access or reset the value of .lambda0.


Create a new GenEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$new(n_obs, n_eqs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_obs}}{Number of observations.}

\item{\code{n_eqs}}{Number of (moment constraint) equations.}

\item{\code{G}}{A numeric matrix.}

\item{\code{value}}{Missing or a positive integer (default to 200).}

\item{\code{value}}{Missing or a small positive number (default to 1e-7).}

\item{\code{value}}{Missing or a vector of length \code{n_eqs}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_supp_adj"></a>}}
\if{latex}{\out{\hypertarget{method-set_supp_adj}{}}}
\subsection{Method \code{set_supp_adj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_supp_adj(supp_adj = FALSE, supp_adj_a = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_opts"></a>}}
\if{latex}{\out{\hypertarget{method-set_opts}{}}}
\subsection{Method \code{set_opts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_opts(
  max_iter = 100,
  rel_tol = 1e-07,
  lambda0 = rep(0, GenEL_get_n_eqs(private$.GEL)),
  supp_adj = FALSE,
  a = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda_nr"></a>}}
\if{latex}{\out{\hypertarget{method-lambda_nr}{}}}
\subsection{Method \code{lambda_nr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$lambda_nr(G, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega_hat"></a>}}
\if{latex}{\out{\hypertarget{method-omega_hat}{}}}
\subsection{Method \code{omega_hat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$omega_hat(G, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel(G, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel_grad"></a>}}
\if{latex}{\out{\hypertarget{method-logel_grad}{}}}
\subsection{Method \code{logel_grad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel_grad(G, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
