% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GenEL.R
\name{GenEL}
\alias{GenEL}
\title{R6 class for EL with general moment specification.}
\description{
R6 class for EL with general moment specification.

R6 class for EL with general moment specification.
}
\details{
A general EL object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenEL$new()}}
\item \href{#method-set_opts}{\code{GenEL$set_opts()}}
\item \href{#method-logel}{\code{GenEL$logel()}}
\item \href{#method-lambda_nr}{\code{GenEL$lambda_nr()}}
\item \href{#method-clone}{\code{GenEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Check the dimension of G is what is expected.


Access or reset the value of .G.


Create a new GenEL object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$new(n_obs, n_eqs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{A numeric matrix.}

\item{\code{value}}{Missing or a numeric matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_opts"></a>}}
\if{latex}{\out{\hypertarget{method-set_opts}{}}}
\subsection{Method \code{set_opts()}}{
Set options for EL evaluation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_opts(max_iter = 100, rel_tol = 1e-07, supp_corr = TRUE, lambda0 = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iter}}{Integer scalar for maximum number of Newton-Raphson steps.}

\item{\code{rel_tol}}{Positive scalar for relative tolerance of Newton-Raphson steps.}

\item{\code{lambda0}}{A numeric vector of length \code{n_eqs} as the initial value of lambda.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel(G, return_omega = FALSE, return_dldG = FALSE, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda_nr"></a>}}
\if{latex}{\out{\hypertarget{method-lambda_nr}{}}}
\subsection{Method \code{lambda_nr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$lambda_nr(G, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
