% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GenEL.R
\name{GenEL}
\alias{GenEL}
\title{R6 Class for general EL models.}
\description{
R6 Class for general EL models.

R6 Class for general EL models.
}
\note{
The returned \code{externalptr} is pointing to the same C++ object as is `private$.gelptr$.  So, modifying one modifies the other.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_obs}}{Number of observations.}

\item{\code{n_eqs}}{Number of observations.}

\item{\code{max_iter}}{Maximum number of Newton-Raphson iterations.}

\item{\code{rel_tol}}{Relative tolerance for Newton-Raphson algorithm.}

\item{\code{lambda0}}{Initialization vector for the Newton-Raphson algorithm.}

\item{\code{supp_adj}}{Whether or not to perform support adjustment.}

\item{\code{supp_adj_a}}{Support adjustment factor.}

\item{\code{weight_adj}}{Weights of pseudo-observation for support adjustment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenEL-new}{\code{GenEL$new()}}
\item \href{#method-GenEL-set_opts}{\code{GenEL$set_opts()}}
\item \href{#method-GenEL-lambda_nr}{\code{GenEL$lambda_nr()}}
\item \href{#method-GenEL-omega_hat}{\code{GenEL$omega_hat()}}
\item \href{#method-GenEL-logel}{\code{GenEL$logel()}}
\item \href{#method-GenEL-logel_grad}{\code{GenEL$logel_grad()}}
\item \href{#method-GenEL-clone}{\code{GenEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-new}{}}}
\subsection{Method \code{new()}}{
Check the dimension of G is what is expected.


Check the dimension of weights is as expected.


Get \code{private$.gelptr}.


Create a new \code{GenEL} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$new(
  n_obs,
  n_eqs,
  max_iter = 100,
  rel_tol = 1e-07,
  supp_adj = FALSE,
  supp_adj_a = max(1, 0.5 * log(n_obs)),
  weight_adj = 1,
  lambda0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_obs}}{Number of observations.}

\item{\code{n_eqs}}{Number of (moment constraint) equations.}

\item{\code{max_iter}}{Maximum number of Newton-Raphson iterations (positive integer).}

\item{\code{max_iter}}{Maximum number of Newton-Raphson iterations (positive integer).}

\item{\code{rel_tol}}{Relative tolerance for Newton-Raphson termination (positive scalar).}

\item{\code{rel_tol}}{Relative tolerance for Newton-Raphson termination (positive scalar).}

\item{\code{supp_adj}}{Whether or not to perform support adjustment (boolean).}

\item{\code{supp_adj}}{Whether or not to perform support adjustment (boolean).}

\item{\code{supp_adj_a}}{Support adjustment factor (positive scalar).}

\item{\code{supp_adj_a}}{Support adjustment factor (positive scalar).}

\item{\code{weight_adj}}{Weight of pseudo-observation when there is support adjustment (positive scalar).}

\item{\code{weight_adj}}{Weight of pseudo-observation when there is support adjustment (positive scalar).}

\item{\code{lambda0}}{Vector of length \code{n_eqs} with which to initialize the Newton-Raphson algorithm.}

\item{\code{lambda0}}{Vector of length \code{n_eqs} with which to initialize the Newton-Raphson algorithm.}

\item{\code{weights}}{A numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This is used to treat \code{.gelptr} as a protected member, i.e., for classes inheriting from \code{GenEL} to pass \code{.gelptr} on to C++ code.  Accessor is private as the intent is for this to not be called directly by the user.
}

\subsection{Returns}{
A \code{GenEL} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-set_opts"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-set_opts}{}}}
\subsection{Method \code{set_opts()}}{
Set multiple options together.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_opts(max_iter, rel_tol, supp_adj, supp_adj_a, weight_adj, lambda0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iter}}{Maximum number of Newton-Raphson iterations (positive integer).}

\item{\code{max_iter}}{Maximum number of Newton-Raphson iterations (positive integer).}

\item{\code{rel_tol}}{Relative tolerance for Newton-Raphson termination (positive scalar).}

\item{\code{rel_tol}}{Relative tolerance for Newton-Raphson termination (positive scalar).}

\item{\code{supp_adj}}{Whether or not to perform support adjustment (boolean).}

\item{\code{supp_adj}}{Whether or not to perform support adjustment (boolean).}

\item{\code{supp_adj_a}}{Support adjustment factor (positive scalar).}

\item{\code{supp_adj_a}}{Support adjustment factor (positive scalar).}

\item{\code{weight_adj}}{Weight of pseudo-observation when there is support adjustment (positive scalar).}

\item{\code{weight_adj}}{Weight of pseudo-observation when there is support adjustment (positive scalar).}

\item{\code{lambda0}}{Vector of length \code{n_eqs} with which to initialize the Newton-Raphson algorithm.}

\item{\code{lambda0}}{Vector of length \code{n_eqs} with which to initialize the Newton-Raphson algorithm.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Any number of these arguments can be supplied.  Those that are missing are not modified.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-lambda_nr"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-lambda_nr}{}}}
\subsection{Method \code{lambda_nr()}}{
Calculate the solution of the empirical likelihood dual problem using the Newton-Raphson algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$lambda_nr(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within \code{max_iter} Newton-Raphson iterations.  If not, returns a vector of \code{NaN}s.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within \code{max_iter} Newton-Raphson iterations.  If not, returns a vector of \code{NaN}s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of length \code{n_eqs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-omega_hat"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-omega_hat}{}}}
\subsection{Method \code{omega_hat()}}{
Calculate the probability vector base on the given G matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$omega_hat(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within \code{max_iter} Newton-Raphson iterations.  If not, returns a vector of \code{NaN}s.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within \code{max_iter} Newton-Raphson iterations.  If not, returns a vector of \code{NaN}s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A probability vector of length \code{n_obs + supp_adj}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-logel"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-logel}{}}}
\subsection{Method \code{logel()}}{
Evaluate the empirical loglikelihood function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within \code{max_iter} Newton-Raphson iterations.  If not, returns `-Inf.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The empirical loglikelihood evaluated at \code{G} (a scalar).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-logel_grad"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-logel_grad}{}}}
\subsection{Method \code{logel_grad()}}{
Calculate empirical loglikelihood and its derivative with respect to \code{G}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel_grad(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{G}}{Constraint matrix of size \verb{n_obs x n_eqs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{weights}}{Non-negative weight of each observation.  A numeric vector of length \code{n_obs}.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within \code{max_iter} Newton-Raphson iterations.  If not, sets the empirical loglikelihood to \code{-Inf} and its gradient to a matrix of \code{NaN}s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements \code{logel} and \code{grad}, the former being a scalaer and the latter a matrix of size \verb{n_obs x n_eqs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenEL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenEL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
