% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_GenEL.R
\name{GenEL}
\alias{GenEL}
\title{General EL class}
\description{
R6 class for EL with general moment specification.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{max_iter}}{Access or reset the value of the maximum number of
iterations. The value must be a positive integer (default to 100).}

\item{\code{rel_tol}}{Access or reset the value of relative tolerance
controlling the accuracy at convergence. The value must be a small
positive number (default to 1e-7).}

\item{\code{lambda0}}{Access or reset the initial value of lambda. The value
must be a vector of length \code{n_eqs} (default to a vector of 0).}

\item{\code{supp_adj}}{Access or reset the support correction flag. The value
must be a boolean indicating whether to conduct support correction or
not (default to FALSE).}

\item{\code{supp_adj_a}}{Access or reset the value of support corection factor.
The value must be a scalar (defaults to \code{max(1.0, log(n_obs)/2)}).}

\item{\code{weight_adj}}{Access or reset the value of the weight for the
additional fake observation under support correction. The value must
be a scalar (default to NULL).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenEL$new()}}
\item \href{#method-set_supp_adj}{\code{GenEL$set_supp_adj()}}
\item \href{#method-set_opts}{\code{GenEL$set_opts()}}
\item \href{#method-lambda_nr}{\code{GenEL$lambda_nr()}}
\item \href{#method-omega_hat}{\code{GenEL$omega_hat()}}
\item \href{#method-logel}{\code{GenEL$logel()}}
\item \href{#method-logel_grad}{\code{GenEL$logel_grad()}}
\item \href{#method-clone}{\code{GenEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Check the dimension of G is what is expected.


Check the dimension of weights is as expected.


Create a new \code{GenEL} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$new(n_obs, n_eqs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_obs}}{Number of observations.}

\item{\code{n_eqs}}{Number of (moment constraint) equations.}

\item{\code{weights}}{A numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{GenEL} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_supp_adj"></a>}}
\if{latex}{\out{\hypertarget{method-set_supp_adj}{}}}
\subsection{Method \code{set_supp_adj()}}{
Set the support correction flag and support correction factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_supp_adj(supp_adj = FALSE, supp_adj_a = NULL, weight_adj = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{supp_adj}}{A boolean indicating whether to conduct support correction or not.}

\item{\code{supp_adj_a}}{Support adjustment factor (default to \code{max(1.0, log(n_obs)/2)}).}

\item{\code{weight_adj}}{Weight under weighted log EL (default to 1.0).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_opts"></a>}}
\if{latex}{\out{\hypertarget{method-set_opts}{}}}
\subsection{Method \code{set_opts()}}{
Set more than one options together.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$set_opts(
  max_iter = 100,
  rel_tol = 1e-07,
  supp_adj = FALSE,
  supp_adj_a = NULL,
  weight_adj = NULL,
  lambda0 = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_iter}}{A positive integer controlling the maximum number of iterations.}

\item{\code{rel_tol}}{A small positive number controlling accuracy at convergence.}

\item{\code{supp_adj}}{A boolean indicating whether to conduct support correction or not.}

\item{\code{supp_adj_a}}{Support adjustment factor. Defaults to \code{max(1.0, log(n_obs)/2)}.}

\item{\code{weight_adj}}{Weight under weighted log EL (default to 1.0).}

\item{\code{lambda0}}{Initialization vector of size \code{n_eqs}.  Defaults to a vector of zeros.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda_nr"></a>}}
\if{latex}{\out{\hypertarget{method-lambda_nr}{}}}
\subsection{Method \code{lambda_nr()}}{
Calculate the solution of the dual problem of maximum log EL problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$lambda_nr(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{A numeric matrix of dimension \verb{n_obs x n_eqs}.}

\item{\code{weights}}{A numeric vector of length \code{n_obs} containing non-negative values.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within the given maximum number of Newton-Raphson iterations.  If not, returns a vector of \code{NaN}s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of length \code{n_eqs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega_hat"></a>}}
\if{latex}{\out{\hypertarget{method-omega_hat}{}}}
\subsection{Method \code{omega_hat()}}{
Calculate the probability vector base on the given G matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$omega_hat(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{A numeric matrix of dimension \verb{n_obs x n_eqs}.}

\item{\code{weights}}{A numeric vector of length \code{n_obs} containing non-negative values.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within the given maximum number of Newton-Raphson iterations.  If not, returns a vector of \code{NaN}s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A probability vector of length \code{n_obs + supp_adj}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
Calculate the log empirical likelihood for a given \code{G} matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{A numeric matrix of dimension \verb{n_obs x n_eqs}.}

\item{\code{weights}}{A numeric vector of length \code{n_obs} containing non-negative values.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within the given maximum number of Newton-Raphson iterations.  If not, returns \code{-Inf}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scalar.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel_grad"></a>}}
\if{latex}{\out{\hypertarget{method-logel_grad}{}}}
\subsection{Method \code{logel_grad()}}{
Calculate log EL and the derivative of log EL w.r.t. G evaluated at G.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$logel_grad(G, weights, check_conv = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{G}}{A numeric matrix of dimension \verb{n_obs x n_eqs}.}

\item{\code{weights}}{A numeric vector of length \code{n_obs} containing non-negative values.}

\item{\code{check_conv}}{If \code{TRUE}, checks whether the desired \code{rel_tol} was reached within the given maximum number of Newton-Raphson iterations.  If not, sets the log EL to negative infinity and the gradient to a matrix of \code{NaN}s.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements \code{logel} and \code{grad}, the latter being a matrix of size \verb{n_obs x n_eqs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
