% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaNR.R
\name{lambdaNR}
\alias{lambdaNR}
\title{Solve the inner loop optimization of an EL function.}
\usage{
lambdaNR(G, lambda0, nObs, nEqs, max_iter = 100, eps = 1e-07,
  verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqns} matrix of constraints.}

\item{lambda0}{Length-\code{nEq} vector of starting values.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{eps}{Relative tolerance of Newton-Raphson convergence.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Length-\code{nEq} vector corresponding to the solution of the optimization problem.
}
\description{
Solve the inner loop optimization of an EL function.
}
\details{
The inner-loop optimization of EL is ...
}
