% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaNR.R
\name{lambdaNR}
\alias{lambdaNR}
\title{Solve the inner loop optimization of an EL function.}
\usage{
lambdaNR(G, ws, max_iter = 100, rel_tol = 1e-07, verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqns} matrix of constraints.}

\item{ws}{Optional length-\code{nObs} vector of weights for weighted Newton-Raphson algorithm.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Length-\code{nEq} vector corresponding to the solution of the optimization problem.
}
\description{
Solve the inner loop optimization of an EL function.
}
\details{
The inner-loop optimization of EL is ...
}
