% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{Calculates log empirical likelihood.}
\usage{
logEL(
  G,
  lambda0 = rep(0, ncol(G)),
  delta = NULL,
  eps = NULL,
  support = FALSE,
  sp = 0,
  max_iter = 100,
  rel_tol = 1e-07,
  abs_tol = 0.001,
  return_omega = FALSE,
  return_dldG = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{G}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}

\item{lambda0}{A numeric vector of length \code{n_eqs} as the initial value of lambda.}

\item{delta}{Length-\code{nObs} vector of censoring indicators (1 right-censord, 0 not-censored).}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{support}{A boolean indicating whether support correction should be conducted.}

\item{sp}{A positive scalar as smoothing parameter.}

\item{max_iter}{Integer scalar for maximum number of Newton-Raphson steps.}

\item{rel_tol}{Positive scalar for relative tolerance of Newton-Raphson steps.}

\item{abs_tol}{Positive scalar for absolute tolerance of Newton-Raphson steps.}

\item{return_omega}{A boolean indicating whether to return the probability vector omega used to evaluate log EL.}

\item{return_dldG}{A boolean indicating whether to return the gradient matrix dldG of log EL w.r.t. G.}

\item{verbose}{Logical scalar for whether to display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
A list containing the log EL value, the probability vector omega, and a matrix
containing the gradient of log EL function with respect to G matrix evaluated at G,
or only the log EL value as a scalar.
}
\description{
Calculates log empirical likelihood.
}
\examples{
# no censoring
n <- 20
p <- 5
max_iter <- 200
rel_tol <- 1e-4
G <- matrix(rnorm(n*p),n,p) # random G here
flexEL::logEL(G = G, support = FALSE,
              max_iter = max_iter, rel_tol = rel_tol, abs_tol = 1e-3,
              return_omega = FALSE, verbose = FALSE)

# no censoring, get the derivative of log EL w.r.t. G
flexEL::logEL(G = G, support = FALSE,
              max_iter = max_iter, rel_tol = rel_tol, abs_tol = 1e-3,
              return_omega = FALSE, return_dldG = TRUE, verbose = FALSE)$dldG

# right censoring
n <- 20
p <- 5
max_iter <- 200
rel_tol <- 1e-4
abs_tol <- 1e-3
G <- matrix(rnorm(n*p), n, p)
deltas <- rep(1,n)
numcens <- sample(round(n/2),1)
censinds <- sample(n,numcens)
deltas[censinds] <- 0
epsilons <- rnorm(n)
flexEL::logEL(G = G, delta = deltas, eps = epsilons, support = FALSE, 
              max_iter = max_iter, rel_tol = rel_tol, abs_tol = abs_tol, verbose = FALSE)
}
