% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{log empirical likelihood}
\usage{
logEL(G, deltas, epsilons, max_iter = 100, rel_tol = 1e-07,
  verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqs} matrix of constraints.}

\item{deltas}{\code{nObs} vector of censoring indicators.}

\item{epsilons}{\code{nObs} vector of residuals.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{omegas}{\code{nObs} probability vector.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood
}
\details{
...
}
