% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{Calculates log empirical likelihood.}
\usage{
logEL(
  G,
  delta = NULL,
  eps = NULL,
  support = FALSE,
  max_iter = 100,
  rel_tol = 1e-07,
  abs_tol = 0.001,
  return_omega = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{G}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}

\item{delta}{Length-\code{nObs} vector of censoring indicators (1 right-censord, 0 not-censored).}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{support}{A boolean indicating whether support correction should be conducted.}
}
\value{
A list containing the log EL value and the probability vector omega, or only the log EL value as a scalar.
}
\description{
Calculates log empirical likelihood.
}
\examples{
# no censoring
n <- 20
omegas <- abs(rnorm(n))
omegas <- omegas/sum(omegas) # create a random probability vector
logEL(omegas)

# right censoring
n <- 20 # total number of observations
k <- 4 # number of censored observations
idx <- sample(n,k) # indices of censored observations
deltas <- rep(1,n)
deltas[idx] <- 0 # censoring indicators
omegas <- abs(rnorm(n)) 
omegas <- omegas/sum(omegas) # create a random probability vector
epsilons <- rnorm(n) # time to event for each observation
logEL(omegas,epsilons,deltas)
}
