% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{log empirical likelihood}
\usage{
logEL(omega, eps, delta, support = FALSE)
}
\arguments{
\item{omega}{Length-\code{nObs} probability vector.}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{delta}{Length-\code{nObs} vector of censoring indicators.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood
}
\details{
...
}
