% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{log empirical likelihood}
\usage{
logEL(omegas, epsilons, deltas)
}
\arguments{
\item{omegas}{\code{nObs} vector of empirical distribution.}

\item{epsilons}{\code{nObs} vector of residuals.}

\item{deltas}{\code{nObs} vector of censoring indicators.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
log empirical likelihood
}
\details{
...
}
