% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.R
\name{logEL}
\alias{logEL}
\title{Calculate the log empirical likelihood for a given \code{G} matrix.}
\usage{
logEL(G, max_iter = 100, rel_tol = 1e-07, supp_adj = FALSE, grad = FALSE)
}
\arguments{
\item{G}{A numeric matrix of size \verb{n_obs x n_eqs}.}

\item{max_iter}{A positive integer controlling the maximum number of iterations.}

\item{rel_tol}{A small positive number controlling accuracy at convergence.}

\item{supp_adj}{A boolean indicating whether to conduct support correction or not.}

\item{grad}{Whether or not to return the gradient of the log empirical likelihood with respect to \code{G}.}
}
\value{
The log empirical likelihood evaluated at \code{G}, or if \code{grad == TRUE}, a list with elements \code{logel} and \code{grad}, with the latter being a matrix of size \verb{n_obs x n_eqs}.
}
\description{
A simplified wrapper to \code{\link[=GenEL$logEL]{GenEL$logEL()}} and \code{\link[=GenEL$logEL_grad]{GenEL$logEL_grad()}}.  See documentation therein for more options.
}
