% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL.smooth.R
\name{logEL.smooth}
\alias{logEL.smooth}
\title{Calculate the smoothed log empirical likelihood under censoring}
\usage{
logEL.smooth(omega, eps, deltas, sp = 10, support = FALSE)
}
\arguments{
\item{omega}{Length-\code{nObs} probability vector.}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{sp}{Positive scalar as smoothing parameter.}
}
\value{
Smoothed log empirical likelihood (a scalar).
}
\description{
Calculate the smoothed log empirical likelihood under censoring
}
\details{
...
}
