% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEL_smooth.R
\name{logEL_smooth}
\alias{logEL_smooth}
\title{Calculate log empirical likelihood under censoring with continuity correction.}
\usage{
logEL_smooth(omega, eps, deltas, sp = 10, support = FALSE)
}
\arguments{
\item{omega}{Length-\code{nObs} probability vector.}

\item{eps}{Length-\code{nObs} vector of residuals.}

\item{sp}{Positive scalar as smoothing parameter.}
}
\value{
Log empirical likelihood with continuity correction.
}
\description{
Calculate log empirical likelihood under censoring with continuity correction.
}
\examples{
n <- 20 # total number of observations
s <- 10 # parameter for continuity correction
omegas <- abs(rnorm(n))
omegas <- omegas/sum(omegas) # a probability vector
epsilons <- rnorm(n)
deltas <- rep(1,n)
numcens <- 5
deltas[sample(1:n,numcens)] <- 0 # censoring indicators
logEL_smooth(omegas,epsilons,deltas,s)
}
