% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MeanRegEL.R
\name{MeanRegEL}
\alias{MeanRegEL}
\title{Mean regression EL class}
\description{
R6 class for EL with mean regression moment specification.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeanRegEL$new()}}
\item \href{#method-eval_G}{\code{MeanRegEL$eval_G()}}
\item \href{#method-logel}{\code{MeanRegEL$logel()}}
\item \href{#method-logel_grad}{\code{MeanRegEL$logel_grad()}}
\item \href{#method-get_el_obj}{\code{MeanRegEL$get_el_obj()}}
\item \href{#method-clone}{\code{MeanRegEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanRegEL$new(y, X, Z = NULL, delta = NULL, el_opts = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_G"></a>}}
\if{latex}{\out{\hypertarget{method-eval_G}{}}}
\subsection{Method \code{eval_G()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanRegEL$eval_G(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanRegEL$logel(theta, grad_out = TRUE, negate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel_grad"></a>}}
\if{latex}{\out{\hypertarget{method-logel_grad}{}}}
\subsection{Method \code{logel_grad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanRegEL$logel_grad(theta, logel_out = FALSE, negate = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_el_obj"></a>}}
\if{latex}{\out{\hypertarget{method-get_el_obj}{}}}
\subsection{Method \code{get_el_obj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanRegEL$get_el_obj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanRegEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
