% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.evalG.R
\name{mr.evalG}
\alias{mr.evalG}
\title{Solve the inner loop optimization of an EL function.}
\usage{
mr.evalG(y, X, beta)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nEqns} matrix of covariates.}

\item{beta}{Length-\code{nEqs} vector of coefficients in location model.}
}
\value{
G \code{nObs x nEqns} matrix of constraints.
}
\description{
Solve the inner loop optimization of an EL function.
}
\details{
Returns the G matrix for location mean regression model.
}
