% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_cens.neglogEL.smooth.R
\name{mr_cens.neglogEL.smooth}
\alias{mr_cens.neglogEL.smooth}
\title{Calculate the negative log EL with smoothing for mean regression under censoring}
\usage{
mr_cens.neglogEL.smooth(y, X, deltas, beta, s = 10)
}
\arguments{
\item{omegas}{Vector of probability weights}

\item{epsilons}{Vector of residuals.}

\item{ii}{An integer indicator between 1 and the length of omegas/epsilons.}
}
\value{
Partial sum of omegas according to residuals that are no larger than the ii-th residual.
}
\description{
Calculate the negative log EL with smoothing for mean regression under censoring
}
\details{
...
}
