% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls.post.R
\name{mrls.post}
\alias{mrls.post}
\title{Posterior sampler for mean regression (location-scale model)}
\usage{
mrls.post(y, X, Z, nsamples, nburn, BetaInit, GammaInit, Sig2Init, mwgSd,
  max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nBet} matrix of location function covariates.}

\item{Z}{\code{nObs x nGam} matrix of scale function covariates.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{BetaInit}{\code{nBet x numTheta} matrix of initial value for the parameters in the linear location function.}

\item{GammaInit}{\code{nGam x numTheta} matrix of initial value for the parameters in the exponential scale function.}

\item{Sig2Init}{\code{numTheta} vector of initial value for the squared scale parameter.}

\item{mwgSd}{Length-\code{nEqs} vector of tuning parameters.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for mean regression (location-scale model)
}
\details{
...
}
