% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls_evalG.R
\name{mrls_evalG}
\alias{mrls_evalG}
\title{Evaluate the G matrix for a location-scale mean regression model.}
\usage{
mrls_evalG(y, X, Z, beta, gamma, sig2)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates in the location function, where \code{nObs} is the number of observations and \code{nBet} is the dimension of coefficients in the location function.}

\item{Z}{\code{nObs x nGam} matrix of covariates in the scale function, where \code{nObs} is the number of observations and \code{nGam} is the dimension of coefficients in the scale function.}

\item{beta}{Length-\code{n_bet} vector of coefficients in location model.}

\item{gamma}{Length-\code{n_gam} vector of coefficients in location model.}

\item{sig2}{A positive scalar whose square root is the scale parameter for the error term.}
}
\value{
G matrix for location-scale model.
}
\description{
Evaluate the G matrix for a location-scale mean regression model.
}
\details{
Assuming data were generated from\preformatted{y_i = x_i'beta + sigma * exp(z_i'gamma) * eps_i, for i = 1, ..., n,
}

where \code{eps_i}'s are ~iid \code{eps}, with \code{E[eps] = 0} and \code{Var[eps] = 1}.
}
\examples{
n <- 20
p <- 5
q <- 3
X <- replicate(p, rnorm(n))
Z <- replicate(q, rnorm(n))
beta <- rnorm(p)
gamma <- rnorm(q)
sig2 <- 1.5
y <- c(X \%*\% beta + exp(Z \%*\% gamma)) + rnorm(n) # with N(0,1) error term
max_iter <- 100
rel_tol <- 1e-5
mrls_evalG(y,X,Z,beta,gamma,sig2)
}
