% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrls_evalG.R
\name{mrls_evalG}
\alias{mrls_evalG}
\title{Evaluate the G matrix for a location-scale mean regression model.}
\usage{
mrls_evalG(y, X, Z, beta, gamma, sig2)
}
\arguments{
\item{y}{A numeric vector of resoponses of length \code{n_obs} where \code{n_obs}
is the number of observations.}

\item{X}{A numeric matrix of covariates in the location function of dimension
\code{n_obs} x \code{n_bet} where \code{n_obs} is the number of observations and
\code{b_bet} is the number of location function's coefficients (length of \code{beta}).}

\item{Z}{A numeric matrix of covariates in the scale function of dimension
\code{n-obs} x \code{n_gam} where \code{n_obs} is the number of observations and
\code{b_gam} is the number of scale function's coefficients (length of \code{gamma}).}

\item{beta}{A numeric vector of coefficients in the location function.}

\item{gamma}{A numeric vector of coefficients in the scale function.}

\item{sig2}{A positive scalar whose square root is the scale parameter in the scale function.}
}
\value{
A numeric matrix of dimension \code{n_obs} x \code{n_bet + n_gam + 1}.
}
\description{
Evaluate the G matrix for a location-scale mean regression model.
}
\details{
Assuming data were generated from\preformatted{y_i = x_i'beta + sqrt(sig2) * exp(z_i'gamma) * eps_i, for i = 1, ..., n,
}

where \code{eps_i}'s are ~iid \code{F(eps)}, with \code{E[eps] = 0} and \code{Var[eps] = 1}.
In this location-scale model, \verb{x_i'beta} is the location function and \verb{sqrt(sig2) * exp(z_i'gamma)}
is the scale function. The \code{G} matrix is calculated by stacking together the first derivative of the
quasi-likelihood function w.r.t \code{beta}, \code{gamma}, and \code{sig2}.
More details of this setup can be found in the package vignette:
\code{vignette("help", package = "flexEL")}.
}
\examples{
## simulate some data ##
n <- 20
p <- 2
q <- 2
X <- replicate(p-1, rnorm(n))
X <- cbind(1, X) # X may have an intercept term, if so, it should be explicit
Z <- replicate(q, rnorm(n)) # Z shall not have an intercept term
beta0 <- c(1, 2)
gamma0 <- c(0.5, 0.25)
sig20 <- 0.5
y <- c(X \%*\% beta0 + sqrt(sig20)*exp(Z \%*\% gamma0)*rnorm(n)) # with N(0,1) error term

## calculate G matrix given data and certain parameter values ##
beta <- c(1,2)
gamma <- c(0.5, 0.25)
sig2 <- 0.5
mrls_evalG(y,X,Z,beta,gamma,sig2)
}
