% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.hat.R
\name{omega.hat}
\alias{omega.hat}
\title{Returns the empirical distribution, omegas, from empirical likelihood (EL) maximization.}
\usage{
omega.hat(G, deltas, epsilons, max_iter = 100, rel_tol = 1e-07,
  verbose = FALSE)
}
\arguments{
\item{G}{\code{nObs x nEqns} matrix of constraints.}

\item{epsilons}{Length-\code{nObs} vector of residuals, omit if no censoring.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}

\item{delta}{Length-\code{nObs} vector of censor indicators, omit if no censoring.}
}
\value{
Length-\code{nEq} vector for the resulting empirical distribution, omegas.
}
\description{
Returns the empirical distribution, omegas, from empirical likelihood (EL) maximization.
}
\details{
The inner-loop optimization of EL is ...
}
