% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_hat.R
\name{omega_hat}
\alias{omega_hat}
\title{Returns the empirical distribution, omegas, from empirical likelihood (EL) maximization.}
\usage{
omega_hat(G, deltas, epsilons, max_iter = 100, rel_tol = 1e-07,
  abs_tol = 0.001, support = FALSE, verbose = FALSE)
}
\arguments{
\item{G}{\code{n_obs x n_eqs} matrix of constraints.}

\item{deltas}{Length-\code{n_obs} vector of censor indicators, omit if no censoring.}

\item{epsilons}{Length-\code{n_obs} vector of residuals, omit if no censoring.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{verbose}{Display number of steps and tolerance criterion when algorithm terminates.}
}
\value{
Length-\code{n_eqs} vector for the resulting empirical distribution, omegas, if the optimization algorithm converged; 1/n_obs if did not converge.
}
\description{
Returns the empirical distribution, omegas, from empirical likelihood (EL) maximization.
}
\examples{
# no censoring
n <- 20
p <- 7
G <- matrix(rnorm(n*p),n,p) # random G here
max_iter <- 100
rel_tol <- 1e-5
omega_hat(G = G, max_iter = max_iter, rel_tol = rel_tol, 
          support = FALSE, verbose = FALSE) # without support correction
omega_hat(G = G, max_iter = max_iter, rel_tol = rel_tol, 
          support = TRUE, verbose = FALSE) # with support correction

# right-censoring
n <- 20
p <- 7
G <- matrix(rnorm(n*p),n,p) # random G here
max_iter <- 100
rel_tol <- 1e-5
abs_tol <- 1e-3
deltas <- rep(1,n)
numcens <- 5
censinds <- sample(n,numcens)
deltas[censinds] <- 0
epsilons <- rnorm(n)
omega_hat(G, deltas, epsilons, max_iter = max_iter, 
          rel_tol = rel_tol, abs_tol = abs_tol, support = FALSE, verbose = FALSE) # without support correction
omega_hat(G, deltas, epsilons, max_iter = max_iter, 
          rel_tol = rel_tol, abs_tol = abs_tol, support = TRUE, verbose = FALSE) # with support correction
}
