% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.evalG.R
\name{qr.evalG}
\alias{qr.evalG}
\title{Evaluate the G matrix for quantile regression (location model)}
\usage{
\method{qr}{evalG}(y, X, alpha, beta)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nEqns} matrix of covariates.}

\item{beta}{Length-\code{nEqs} vector of coefficients in location model.}
}
\description{
Evaluate the G matrix for quantile regression (location model)
}
\details{
Returns the G matrix for location mean regression model.
}
