% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.post.R
\name{qr.post}
\alias{qr.post}
\title{Posterior sampler for quantile regression (location model)}
\usage{
\method{qr}{post}(y, X, alpha, nsamples, nburn, betaInit, sigs,
  max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nEqs} matrix of constraints.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}

\item{beta}{Length-\code{nVars} vector of coefficients.}
}
\value{
Log empirical likelihood of the input beta
}
\description{
Posterior sampler for quantile regression (location model)
}
\details{
...
}
