% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.post.R
\name{qr.post}
\alias{qr.post}
\title{Posterior sampler for quantile regression (location model)}
\usage{
\method{qr}{post}(y, X, alpha, nsamples, nburn, betaInit, sigs,
  max_iter = 100, rel_tol = 1e-07)
}
\arguments{
\item{y}{Length-\code{nObs} vector of response values.}

\item{X}{\code{nObs x nEqs} matrix of constraints.}

\item{alpha}{A scalar of quantile level.}

\item{nsamples}{Number of samples to obtain.}

\item{nburn}{number of samples to discard before saving the chain.}

\item{betaInit}{Length-\code{nEqs} vector of initial value for the chain.}

\item{sigs}{Length-\code{nObs} vector of tuning parameters.}

\item{max_iter}{Maximum number of Newton-Raphson steps.}

\item{rel_tol}{Relative tolerance of Newton-Raphson convergence.}
}
\value{
\code{nEqs x nsamples} matrix of Markov Chain.
}
\description{
Posterior sampler for quantile regression (location model)
}
\details{
...
}
