% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_evalG.R
\name{qr_evalG}
\alias{qr_evalG}
\title{Evaluate the G matrix for quantile regression (location model)}
\usage{
qr_evalG(y, X, alphas, Beta)
}
\arguments{
\item{y}{Length-\code{n_obs} vector of responses.}

\item{X}{\code{n_obs x n_eqs} matrix of covariates.}

\item{alphas}{a vector of quantile levels.}

\item{Beta}{\code{n_eqs x n_qts} matrix, each column is a vector of coefficients in location model.}
}
\value{
G matrix for location quantile regression model.
}
\description{
Evaluate the G matrix for quantile regression (location model)
}
\details{
...
}
