% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_evalG.R
\name{qr_evalG}
\alias{qr_evalG}
\title{Evaluate the G matrix for the quantile regression model.}
\usage{
qr_evalG(y, X, alpha, Beta)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates in the location function, where \code{nObs} is the number of observations and \code{nBet} is the dimension of coefficients in the location function.}

\item{alpha}{a vector of quantile levels.}

\item{Beta}{\verb{n_eqs x n_qts} matrix, each column is a vector of coefficients in location model.}
}
\value{
G matrix for location quantile regression model.
}
\description{
Evaluate the G matrix for the quantile regression model.
}
\examples{
n <- 20
p <- 4
X <- replicate(p, rnorm(n))
X[1,] <- rep(1,p)
beta <- rnorm(p)
alpha <- runif(1)
y <- c(X \%*\% beta) + rnorm(n) # with N(0,1) error term
qr_evalG(y, X, alpha, beta)
}
