% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls.evalG.R
\name{qrls.evalG}
\alias{qrls.evalG}
\title{Evaluate the G matrix for mean regression (location-scale model)}
\usage{
qrls.evalG(y, X, Z, alpha, beta, gamma)
}
\arguments{
\item{y}{Length-\code{nObs} vector of responses.}

\item{X}{\code{nObs x nBet} matrix of covariates.}

\item{Z}{\code{nObs x nGam} matrix of covariates.}

\item{alpha}{a scalar of quantile level.}

\item{beta}{Length-\code{nBet} vector of coefficients in location model.}

\item{gamma}{Length-\code{nGam} vector of coefficients in location model.}
}
\value{
G matrix for location-scale quantile regression model.
}
\description{
Evaluate the G matrix for mean regression (location-scale model)
}
