% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls_evalG.R
\name{qrls_evalG}
\alias{qrls_evalG}
\title{Evaluate the G matrix for a location-scale quantile regression model.}
\usage{
qrls_evalG(y, X, Z, alpha, beta, gamma, sig2, nu, s = NULL)
}
\arguments{
\item{y}{A numeric vector of resoponses of length \code{n_obs} where \code{n_obs}
is the number of observations.}

\item{X}{A numeric matrix of covariates in the location function of dimension
\code{n_obs} x \code{n_bet} where \code{n_obs} is the number of observations and
\code{b_bet} is the number of location function's coefficients (length of \code{beta}).}

\item{Z}{A numeric matrix of covariates in the scale function of dimension
\code{n-obs} x \code{n_gam} where \code{n_obs} is the number of observations and
\code{b_gam} is the number of scale function's coefficients (length of \code{gamma}).}

\item{alpha}{A numeric vector of quantile levels of length \code{n_qts}.}

\item{beta}{A numeric matrix of length \code{n_bet}.}

\item{gamma}{A An numeric matrix of dimension \code{n_gam}.}

\item{sig2}{A positive scalar whose square root isthe scale parameter in the scale function.}

\item{nu}{A numeric vector of quantile values of length \code{n_qts} corresponding to each alpha.}

\item{s}{A positive scalar as the smoothing parameter. If \code{s = NULL}, then no smoothing is performed.}
}
\value{
A numeric matrix of dimension \code{n_obs} x \code{n_bet + n_gam + 2}.
}
\description{
Evaluate the G matrix for a location-scale quantile regression model.
}
\details{
Assuming data were generated from\preformatted{y_i = x_i'beta + sigma * exp(z_i'gamma) * eps_i, for i = 1, ..., n,
}

where \code{eps_i}'s are ~iid \code{F(eps)}, with \code{E[eps] = 0} and \code{Var[eps] = 1}.
Quantile regression estimates the alpha-level quantile of the response variable, i.e.,\preformatted{Q_alpha(y | x_i, z_i) = x_i'beta + sigma * exp(z_i'gamma) * nu_alpha,
for i = 1, ..., n
}

where \code{nu_alpha} is the alpha-level quantile value of \code{eps}.
In this location-scale model, \verb{x_i'beta} is the location function and \verb{sqrt(sig2) * exp(z_i'gamma)}
is the scale function. The \code{G} matrix is calculated by stacking together the first derivative of the
quasi-likelihood function w.r.t \code{beta}, \code{gamma}, \code{sig2}, and the first derivative w.r.t \code{nu_alpha}
of the check function introduced by Koenker and Bassett (1978),\preformatted{rho_alpha(u) = u * (alpha - 1\{u <= 0\})
}

where \code{alpha} is the quantile level and \verb{1\{\}} is the indicator function which returns 1 if the
condition is true and 0 otherwise.
More details of this setup can be found in the package vignette:
\code{vignette("help", package = "flexEL")}.
}
\examples{
## simulate some data ##
n <- 20
p <- 2
q <- 2
X <- replicate(p-1, rnorm(n))
X <- cbind(1, X) # X may have an intercept term, if so, it should be explicit
Z <- replicate(q, rnorm(n)) # Z shall not have an intercept term
beta0 <- c(1, 2)
gamma0 <- c(0.5, 0.25)
sig20 <- 0.5
y <- c(X \%*\% beta0 + sqrt(sig20)*exp(Z \%*\% gamma0)*rnorm(n)) # with N(0,1) error term

## calculate G matrix given data and certain parameter values ##
# a single quantile level (with continuity correction)
alpha <- 0.5
beta <- c(1, 2)
gamma <- c(0.5, 0.25)
sig2 <- 0.5
nu <- 0
qrls_evalG(y, X, Z, alpha, beta, gamma, sig2, nu, s = 1)
# multiple quantile levels (with continuity correction)
alpha <- c(0.25, 0.75)
beta <- c(1, 2)
gamma <- c(0.5, 0.25)
sig2 <- 0.5
nu <- c(-0.5, 0.5)
qrls_evalG(y, X, Z, alpha, beta, gamma, sig2, nu, s = 1)
}
\references{
G. Basset and R. Koenker. Regression quantiles. Econometrica, 46(1):33–50, 1978.
}
