% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls_evalG.R
\name{qrls_evalG}
\alias{qrls_evalG}
\title{Evaluate the G matrix for a location-scale quantile regression model.}
\usage{
qrls_evalG(y, X, Z, alpha, Beta, Gamma, Sig2, Nu, sp = 0)
}
\arguments{
\item{y}{A numeric vector of resoponses of length \code{n_obs} where \code{n_obs}
is the number of observations.}

\item{X}{A numeric matrix of covariates in the location function of dimension
\code{n_obs} x \code{n_bet} where \code{n_obs} is the number of observations and
\code{b_bet} is the number of location function's coefficients (length of \code{beta}).}

\item{Z}{A numeric matrix of covariates in the scale function of dimension
\code{n-obs} x \code{n_gam} where \code{n_obs} is the number of observations and
\code{b_gam} is the number of scale function's coefficients (length of \code{gamma}).}

\item{alpha}{A numeric vector of quantile levels of length \code{n_qts}.}

\item{Beta}{A numeric matrix of dimension \code{n_bet} x \code{n_qts}. Each column of \code{Beta}
is a vector of coefficients in the location function.}

\item{Gamma}{A An numeric matrix of dimension \code{n_gam} x \code{n_qts}. Each column of \code{Beta}
is a vector of coefficients in the scale function.}

\item{Sig2}{A positive numeric vector of length \code{n_qts} where each element' square root is
the scale parameter in the scale function corresponding to the quantile level.}

\item{Nu}{A numeric vector of quantile values of length \code{n_qts} corresponding to each alpha.}

\item{sp}{A positive scalar as the smoothing parameter. If \code{sp = 0}, then no smoothing is performed.}
}
\value{
A numeric matrix of dimension \code{n_obs} x \code{n_bet + n_gam + 2}.
}
\description{
Evaluate the G matrix for a location-scale quantile regression model.
}
\details{
Assuming data were generated from\preformatted{y_i = x_i'beta + sigma * exp(z_i'gamma) * eps_i, for i = 1, ..., n,
}

where \code{eps_i}'s are ~iid \code{F(eps)}, with \code{E[eps] = 0} and \code{Var[eps] = 1}.
Quantile regression estimates the alpha-level quantile of the response variable, i.e.,\preformatted{Q_alpha(y | x_i, z_i) = x_i'beta + sigma * exp(z_i'gamma) * nu_alpha,
for i = 1, ..., n
}

where \code{nu_alpha} is the alpha-level quantile value of \code{eps}.
In this location-scale model, \verb{x_i'beta} is the location function and \verb{sqrt(sig2) * exp(z_i'gamma)}
is the scale function. The \code{G} matrix is calculated by stacking together the first derivative of the
quasi-likelihood function w.r.t \code{beta}, \code{gamma}, \code{sig2}, and the first derivative w.r.t \code{nu_alpha}
of the check function introduced by \insertCite{koenker-bassett78;textual}{flexEL},\preformatted{rho_alpha(u) = u * (alpha - 1\{u <= 0\})
}

where \code{alpha} is the quantile level and \verb{1\{\}} is the indicator function which returns 1 if the
condition is true and 0 otherwise.
More details of this setup can be found in the package vignette:
\code{vignette("help", package = "flexEL")}.
}
\examples{
n <- 20
p <- 5
q <- 3
X <- replicate(p, rnorm(n))
Z <- replicate(q, rnorm(n))
alpha <- runif(1)
beta <- rnorm(p)
gamma <- rnorm(q)
nu <- rnorm(1)
sig2 <- 1.5
y <- c(X \%*\% beta + sqrt(sig2)*exp(Z \%*\% gamma)*rnorm(n)) # with multiplicative N(0,1) error

# no continuity correction
qrls_evalG(y, X, Z, alpha, beta, gamma, sig2, nu)

# with continuity correction
sp <- 10
qrls_evalG(y, X, Z, alpha, beta, gamma, sig2, nu, sp)
}
\references{
\insertRef{koenker-bassett78}{flexEL}
}
