% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls_evalG.R
\name{qrls_evalG}
\alias{qrls_evalG}
\title{Evaluate the G matrix for quantile regression location-scale model}
\usage{
qrls_evalG(y, X, Z, alphas, Beta, Gamma, Sig2, Nu)
}
\arguments{
\item{y}{Length-\code{n_obs} vector of responses.}

\item{X}{\code{n_obs x n_bet} matrix of covariates.}

\item{Z}{\code{n_obs x n_gam} matrix of covariates.}

\item{alphas}{a vector of quantile levels.}

\item{Beta}{\code{n_bet x n_qts} matrix, each column is a vector of coefficients in location function.}

\item{Gamma}{\code{n_gam x n_qts} matrix, each column is a vector of coefficients in scale function.}

\item{Nu}{Length-\code{numNu} vector of initial value for the chain.}
}
\value{
G matrix for location-scale quantile regression model.
}
\description{
Evaluate the G matrix for quantile regression location-scale model
}
