% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrls_evalG_smooth.R
\name{qrls_evalG_smooth}
\alias{qrls_evalG_smooth}
\title{Evaluate the G matrix for smoothed quantile regression location-scale model}
\usage{
qrls_evalG_smooth(y, X, Z, alphas, Beta, Gamma, Sig2, Nu, sp = 10)
}
\arguments{
\item{y}{Length-\code{n_obs} vector of responses.}

\item{X}{\code{n_obs x n_bet} matrix of covariates.}

\item{Z}{\code{n_obs x n_gam} matrix of covariates.}

\item{alphas}{a vector of quantile levels.}

\item{Beta}{\code{n_bet x n_qts} matrix, each column is a vector of coefficients in location function.}

\item{Gamma}{\code{n_gam x n_qts} matrix, each column is a vector of coefficients in scale function.}

\item{Nu}{Length-\code{numNu} vector of initial value for the chain.}

\item{s}{A positive scalar as smoothing parameter.}
}
\value{
G matrix for location-scale quantile regression model.
}
\description{
Evaluate the G matrix for smoothed quantile regression location-scale model
}
