% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_QuantRegEL.R
\name{QuantRegEL}
\alias{QuantRegEL}
\title{Quantile regression EL class}
\description{
R6 class for EL with location or location-scale quantile regression moment specification.
}
\seealso{
\code{\link[=GenEL$set_opts]{GenEL$set_opts()}}, \code{\link[=CensEL$set_opts]{CensEL$set_opts()}}

\code{\link[=qr_evalG]{qr_evalG()}}, \code{\link[=qrls_evalG]{qrls_evalG()}}

\code{\link[=GenEL$logel]{GenEL$logel()}}, \code{\link[=CensEL$logel]{CensEL$logel()}},
\code{\link[=GenEL$logel_grad]{GenEL$logel_grad()}}, \code{\link[=CensEL$logel_grad]{CensEL$logel_grad()}}

\code{\link[=GenEL$logel]{GenEL$logel()}}, \code{\link[=CensEL$logel]{CensEL$logel()}},
\code{\link[=GenEL$logel_grad]{GenEL$logel_grad()}}, \code{\link[=CensEL$logel_grad]{CensEL$logel_grad()}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QuantRegEL$new()}}
\item \href{#method-eval_G}{\code{QuantRegEL$eval_G()}}
\item \href{#method-logel}{\code{QuantRegEL$logel()}}
\item \href{#method-logel_grad}{\code{QuantRegEL$logel_grad()}}
\item \href{#method-get_el_obj}{\code{QuantRegEL$get_el_obj()}}
\item \href{#method-clone}{\code{QuantRegEL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{QuantRegEL} object. This class can be used for
either non-censored or right-censored data, and location or location
mean regression models.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantRegEL$new(y, X, Z = NULL, delta = NULL, alpha, sp = 0, el_opts = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{A numeric vector of resoponses of length \code{n_obs} where \code{n_obs}
is the number of observations.}

\item{\code{X}}{A numeric matrix of covariates in the location function of dimension
\code{n_obs} x \code{n_bet} where \code{n_obs} is the number of observations and
\code{n_bet} is the number of location function's coefficients (length of \code{beta}).}

\item{\code{Z}}{A numeric matrix of covariates in the scale function of dimension
\code{n-obs} x \code{n_gam} where \code{n_obs} is the number of observations and
\code{b_gam} is the number of scale function's coefficients (length of \code{gamma}).}

\item{\code{delta}}{Length-\code{n_obs} vector of censoring indicators (1 right-censord, 0 not-censored).}

\item{\code{alpha}}{A numeric vector of quantile levels of length \code{n_qts}.}

\item{\code{sp}}{A positive scalar as the smoothing parameter. If \code{s = NULL}, then no smoothing is performed.}

\item{\code{el_opts}}{A list of options for the \code{GenEL} or \code{CensEL} object.
See \code{set_opts} method in the corresponding class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{QuantRegEL} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_G"></a>}}
\if{latex}{\out{\hypertarget{method-eval_G}{}}}
\subsection{Method \code{eval_G()}}{
Evaluate the G matrix for a given parameter value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantRegEL$eval_G(theta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{Length-\code{n_eqs} vector of parameters for a location or location-scale model. This vector should correspond to \code{beta} for a location regression model, or concatenated \code{beta}, \code{gamma}, and \code{sig2} for a location-scale regression model, in this order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric matrix of dimension \verb{n_obs x n_eqs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel"></a>}}
\if{latex}{\out{\hypertarget{method-logel}{}}}
\subsection{Method \code{logel()}}{
Calculate the log empirical likelihood base on the given
parameter value and in the non-censoring case can also return the
gradient.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantRegEL$logel(theta, grad_out = TRUE, negate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{Length-\code{n_eqs} vector of parameters for a location or location-scale model. This vector should correspond to \code{beta} for a location regression model, or concatenated \code{beta}, \code{gamma}, and \code{sig2} for a location-scale regression model, in this order.}

\item{\code{grad_out}}{If \code{TRUE}, add gradient as an attribute to the return value.}

\item{\code{negate}}{If \code{TRUE}, calculate the negative log likelihood.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scalar with a possible attribute.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logel_grad"></a>}}
\if{latex}{\out{\hypertarget{method-logel_grad}{}}}
\subsection{Method \code{logel_grad()}}{
Calculate the gradient of log empirical likelihood at the
given parameter value for non-censoring data, can also return the
log empirical likelihood.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantRegEL$logel_grad(theta, logel_out = FALSE, negate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{Length-\code{n_eqs} vector of parameters for a location or location-scale model. This vector should correspond to \code{beta} for a location regression model, or concatenated \code{beta}, \code{gamma}, and \code{sig2} for a location-scale regression model, in this order.}

\item{\code{logel_out}}{If \code{TRUE}, add log empirical likelihood as an attribute
to the return value.}

\item{\code{negate}}{If \code{TRUE}, calculate the gradient of negative log likelihood.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_el_obj"></a>}}
\if{latex}{\out{\hypertarget{method-get_el_obj}{}}}
\subsection{Method \code{get_el_obj()}}{
Returns the \code{GenEL} or \code{CensEL} object for inspection or
debugging purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantRegEL$get_el_obj()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{GenEL} or \code{CensEL} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuantRegEL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
